CMS-69 add forms in invoice sections to refund or capture a payment

See merge request stancer/ecommerce-module/prestashop!63

+++ upgrade/upgrade-2.0.3.php
@@ -8,13 +8,18 @@
  *
  * @website   https://www.stancer.com
  */
+require_once _PS_ROOT_DIR_ . '/modules/stancer/install/dbUpgrader.php';
 if (!defined('_PS_VERSION_')) {
     exit;
 }
 
-function upgrade_module_unreleased(Stancer $module): bool
+function upgrade_module_2_0_3(Stancer $module): bool
 {
+    if (!$module->installConfigurations()) {
+        return false;
+    }
+    $module->registerHook('displayAdminOrderSide');
     Configuration::deleteByName('STANCER_AUTH_LIMIT');
 
-    return true;
+    return DbUpgrader::upgradeDbAuthorizeStatus($module);
 }
